introduced base stat on signal class
